grafico de status implementado
