+++ src/estatistica.html
@@ -8,20 +8,16 @@
   <link rel="stylesheet" href="grafico.css">
   <title>Gráficos</title>
 </head>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.9.4/Chart.js"></script>
 
 <body>
 
   <div id="container">
-    <canvas id="myChart"></canvas>
+    <canvas id="myChart" width="400" height="400"></canvas>
   </div>
 
 
-  <footer>
-    <p>Desenvolvido por <a href='https://github.com/dayanersilva' target='_blank'>Dayane Rodrigues</a> e <a
-        href='https://github.com/karolsontak' target='_blank'>Karol Sontak</a></p>
-  </footer>
 
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.7.1/chart.min.js"></script>
   <script src="grafico.js" type="module"></script>
 </body>
