--- src/grafico.css
@@ -1,5 +1,0 @@
-body{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
